ADD: fairy: condensation: Int -> BigInteger

+++ common/src/main/kotlin/miragefairy2024/mod/materials/contents/MinaItem.kt
@@ -1,5 +1,6 @@
 package miragefairy2024.mod.materials.contents
 
 import net.minecraft.world.item.Item
+import java.math.BigInteger
 
-class MinaItem(val mina: Int, settings: Properties) : Item(settings)
+class MinaItem(val mina: BigInteger, settings: Properties) : Item(settings)
